Possible fix for issue 341 : Initializing audio device using Webstart throws ClassNotFoundException
I verified that this works and causes no regressions for standalone applications. Additional testing is needed to see if this fixes the bug while using Java Web Start.

This employs the same fix that was used for a similar SceneGraphIO issue; namely, we now use the default (current) class loader rather than the system class loader to load the class. Only if that fails do we try the system class loader.


git-svn-id: https://svn.java.net/svn/j3d-core-utils~svn/branches/dev-1_5@127 9497e636-51bd-65ba-982d-a4982e1767a5